MUC: Move function to utils
